requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
gunicorn==21.2.0
